refactor: split studio controller and unify host command metadata
